implement real sub-pixel rendering

+++ src/ImageSharp.Drawing/Paths/ShapeRegion.cs
@@ -5,6 +5,7 @@
 
 namespace ImageSharp.Drawing
 {
+    using System;
     using System.Buffers;
     using System.Numerics;
 
@@ -39,7 +40,7 @@ namespace ImageSharp.Drawing
         public override Rectangle Bounds { get; }
 
         /// <inheritdoc/>
-        public override int ScanX(int x, float[] buffer, int length, int offset)
+        public override int ScanX(float x, float[] buffer, int length, int offset)
         {
             Vector2 start = new Vector2(x, this.Bounds.Top - 1);
             Vector2 end = new Vector2(x, this.Bounds.Bottom + 1);
@@ -62,7 +63,7 @@ namespace ImageSharp.Drawing
         }
 
         /// <inheritdoc/>
-        public override int ScanY(int y, float[] buffer, int length, int offset)
+        public override int ScanY(float y, float[] buffer, int length, int offset)
         {
             Vector2 start = new Vector2(this.Bounds.Left - 1, y);
             Vector2 end = new Vector2(this.Bounds.Right + 1, y);
