+++ src/ImageSharp.Drawing/Region.cs
@@ -19,7 +19,7 @@ namespace ImageSharp.Drawing
         /// Gets the bounding box that entirely surrounds this region.
         /// </summary>
         /// <remarks>
-        /// This should always contains all possible points returned from either <see cref="ScanX(int, float[], int, int)"/> or <see cref="ScanY(int, float[], int, int)"/>.
+        /// This should always contains all possible points returned from either <see cref="ScanX(float, float[], int, int)"/> or <see cref="ScanY(float, float[], int, int)"/>.
         /// </remarks>
         public abstract Rectangle Bounds { get; }
 
@@ -31,7 +31,7 @@ namespace ImageSharp.Drawing
         /// <param name="length">The length.</param>
         /// <param name="offset">The offset.</param>
         /// <returns>The number of intersections found.</returns>
-        public abstract int ScanX(int x, float[] buffer, int length, int offset);
+        public abstract int ScanX(float x, float[] buffer, int length, int offset);
 
         /// <summary>
         /// Scans the Y axis for intersections.
@@ -41,6 +41,6 @@ namespace ImageSharp.Drawing
         /// <param name="length">The length.</param>
         /// <param name="offset">The offset.</param>
         /// <returns>The number of intersections found.</returns>
-        public abstract int ScanY(int y, float[] buffer, int length, int offset);
+        public abstract int ScanY(float y, float[] buffer, int length, int offset);
     }
 }
